Migrate submit logs endpoint, support gzip-encoded content for all requests
