+++ server/service/endpoint_osquery.go
@@ -2,7 +2,6 @@ package service
 
 import (
 	"context"
-	"encoding/json"
 
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/go-kit/kit/endpoint"
@@ -33,58 +32,5 @@ func makeEnrollAgentEndpoint(svc fleet.Service) endpoint.Endpoint {
 			return enrollAgentResponse{Err: err}, nil
 		}
 		return enrollAgentResponse{NodeKey: nodeKey}, nil
-	}
-}
-
-////////////////////////////////////////////////////////////////////////////////
-// Submit Logs
-////////////////////////////////////////////////////////////////////////////////
-
-type submitLogsRequest struct {
-	NodeKey string          `json:"node_key"`
-	LogType string          `json:"log_type"`
-	Data    json.RawMessage `json:"data"`
-}
-
-type submitLogsResponse struct {
-	Err error `json:"error,omitempty"`
-}
-
-func (r submitLogsResponse) error() error { return r.Err }
-
-func makeSubmitLogsEndpoint(svc fleet.Service) endpoint.Endpoint {
-	return func(ctx context.Context, request interface{}) (interface{}, error) {
-		req := request.(submitLogsRequest)
-
-		var err error
-		switch req.LogType {
-		case "status":
-			var statuses []json.RawMessage
-			if err := json.Unmarshal(req.Data, &statuses); err != nil {
-				err = osqueryError{message: "unmarshalling status logs: " + err.Error()}
-				break
-			}
-
-			err = svc.SubmitStatusLogs(ctx, statuses)
-			if err != nil {
-				break
-			}
-
-		case "result":
-			var results []json.RawMessage
-			if err := json.Unmarshal(req.Data, &results); err != nil {
-				err = osqueryError{message: "unmarshalling result logs: " + err.Error()}
-				break
-			}
-			err = svc.SubmitResultLogs(ctx, results)
-			if err != nil {
-				break
-			}
-
-		default:
-			err = osqueryError{message: "unknown log type: " + req.LogType}
-		}
-
-		return submitLogsResponse{Err: err}, nil
 	}
 }
