+++ server/service/integration_logger_test.go
@@ -121,7 +121,6 @@ func (s *integrationLoggerTestSuite) TestOsqueryEndpointsLogErrors() {
 	require.Nil(t, err)
 
 	logString := s.buf.String()
-	assert.Contains(t, logString, `{"err":"decoding JSON:`)
 	assert.Contains(t, logString, `invalid character '}' looking for beginning of value","level":"info","path":"/api/v1/osquery/log"}
 `, logString)
 }
